Working on chunks which will be subsets of the group given to a thread
A thread will find the next unfinished group and request a chunk of
tasks from it to execute. It will then execute this tasks and then
repeat the process of finding the next task group to run tasks from.

With a chunk we could look at tiling the execution instead of doing
what's done now where we essentially map a 3-nested loop to threads.
Tiling may give better memory coherency, depending on what we're calling