Make Group.finished private, add fcn to check if all Groups are done
The current Context::current_tasks_done will force the thread to busy
wait on the condition since we provide no condition variable or
something it could block on. But the problem is that we don't know if we
can truly mark the context done until ISPCSync so we might incorrectly
mark it done and then more Groups are pushed. But if these groups
re-mark the context as not finished we'd be ok, since we wouldn't enter
ISPCSync at least until returning from ISPCLaunch.