More work towards thread safety
A context's tasks are now accessed behind an iterator and new tasks
launched through a convenience function to hide the `RwLock` which
manages access to the task list.

I've also switched a Group's next start for the chunk to be behind a
`RwLock` as well so we can see if it has tasks remaining without needing
to fully lock a Mutex and multiple threads can check at once.

This of course comes with the small issue that you could see a group has
tasks left when calling next on `ContextIter` but by the time you
actuall call `chunks` on the group those tasks are finished so you just
get back a None and exit the loop. This is not a problem but does
warrant some comments so it's not assumed you would be able to call
unwrap on chunks.