The syncing thread should sleep if it had no chunks to run
This will prevent it from hammering as much on the context, but scaling
is still very poor. Also it seems that if the number of tasks aren't
divisible by the chunk size some aren't executed properly? Or maybe the
sync isn't waiting correctly? E.g. on aobench I'm seeing some incomplete
and/or black scanlines when running with a chunk size of 5 on the
256x256 image.