Figured out why the ao bench isn't being finished properly
The work imbalance on the last one (getting a single scanline to do)
lets it finish before chunks we handed out previously and it incorrectly
marks the task group as finished. We need to wait for all outstanding
chunks to be finished and start >= end to actually mark ourselves done.