Fix bug with marking groups finished too early
It's wrong to assume that the last chunk to run for a task will be the
last to finish chronologically due to work imbalance. We now track how
many chunks we've launched and how many have finished and use this along
with if we've launched all tasks (start >= end) to determine if a group
is actually finished.