Working on adding point lights, but lots of odd lighting bugs
It's super fucky, unsure why the point light is so odd, it seems like
light dir is wrong and we also can't self-shadow ourselves? So the
backface of the sphere still gets light and also how is the dot product
with the light dir not becoming negative back therE? Makes no sense.