Update to bindgen 0.18.0
This requires a small tweak to the output of bindgen where we move the
`pub mod` string we were placing around the file outside in the
ispc_module macro inside the file. This is done since bindgen will now
generate the scope-wide allow attribute to suppress the warnings about
bad names and such which we where doing previously.