compile: Replace ancient `gcc` crate with `cc`
The `gcc` crate hasn't been updated for 5 years and clearly states that
is deprecated after a rename to `cc`.

More importantly we're cross-compiling some Windows ISPC code to aarch64
and while GitHub's actions runners definitely have [all the cross-
compiling packages installed] the old `gcc` crate was complaining that it
could not find `lib.exe` for the desired target.

Note that this strangely happened to work fine on a developer machine,
but it is unclear why.

[all the cross-compiling packages installed]: https://github.com/actions/runner-images/blob/main/images/windows/Windows2022-Readme.md#workloads-components-and-extensions